+++ envoy/extensions/filters/network/http_connection_manager/v3/http_connection_manager.pb.validate.go
@@ -712,16 +712,6 @@ func (m *Rds) Validate() error {
 
 	// no validation rules for RouteConfigName
 
-	if v, ok := interface{}(m.GetRdsResourceLocator()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return RdsValidationError{
-				field:  "RdsResourceLocator",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
 	return nil
 }