+++ envoy/extensions/filters/network/http_connection_manager/v4alpha/http_connection_manager.pb.validate.go
@@ -700,24 +700,7 @@ func (m *Rds) Validate() error {
 		}
 	}
 
-	switch m.NameSpecifier.(type) {
-
-	case *Rds_RouteConfigName:
-		// no validation rules for RouteConfigName
-
-	case *Rds_RdsResourceLocator:
-
-		if v, ok := interface{}(m.GetRdsResourceLocator()).(interface{ Validate() error }); ok {
-			if err := v.Validate(); err != nil {
-				return RdsValidationError{
-					field:  "RdsResourceLocator",
-					reason: "embedded message failed validation",
-					cause:  err,
-				}
-			}
-		}
-
-	}
+	// no validation rules for RouteConfigName
 
 	return nil
 }