+++ envoy/extensions/filters/network/thrift_proxy/v3/thrift_proxy.pb.validate.go
@@ -90,6 +90,8 @@ func (m *ThriftProxy) Validate() error {
 
 	}
 
+	// no validation rules for PayloadPassthrough
+
 	return nil
 }