+++ envoy/extensions/transport_sockets/tls/v4alpha/secret.pb.validate.go
@@ -119,6 +119,13 @@ func (m *SdsSecretConfig) Validate() error {
 		return nil
 	}
 
+	if utf8.RuneCountInString(m.GetName()) < 1 {
+		return SdsSecretConfigValidationError{
+			field:  "Name",
+			reason: "value length must be at least 1 runes",
+		}
+	}
+
 	if v, ok := interface{}(m.GetSdsConfig()).(interface{ Validate() error }); ok {
 		if err := v.Validate(); err != nil {
 			return SdsSecretConfigValidationError{
@@ -127,31 +134,6 @@ func (m *SdsSecretConfig) Validate() error {
 				cause:  err,
 			}
 		}
-	}
-
-	switch m.NameSpecifier.(type) {
-
-	case *SdsSecretConfig_Name:
-
-		if utf8.RuneCountInString(m.GetName()) < 1 {
-			return SdsSecretConfigValidationError{
-				field:  "Name",
-				reason: "value length must be at least 1 runes",
-			}
-		}
-
-	case *SdsSecretConfig_SdsResourceLocator:
-
-		if v, ok := interface{}(m.GetSdsResourceLocator()).(interface{ Validate() error }); ok {
-			if err := v.Validate(); err != nil {
-				return SdsSecretConfigValidationError{
-					field:  "SdsResourceLocator",
-					reason: "embedded message failed validation",
-					cause:  err,
-				}
-			}
-		}
-
 	}
 
 	return nil