+++ pkg/cache/v2/mux.go
@@ -34,18 +34,19 @@ type MuxCache struct {
 
 var _ Cache = &MuxCache{}
 
-func (mux *MuxCache) CreateWatch(request *Request, value chan<- Response) func() {
+func (mux *MuxCache) CreateWatch(request *Request) (chan Response, func()) {
 	key := mux.Classify(*request)
 	cache, exists := mux.Caches[key]
 	if !exists {
-		value <- nil
-		return nil
+		value := make(chan Response, 0)
+		close(value)
+		return value, nil
 	}
-	return cache.CreateWatch(request, value)
+	return cache.CreateWatch(request)
 }
 
 // TODO: implement CreateDeltaWatch for linear cache
-func (cache *MuxCache) CreateDeltaWatch(request *DeltaRequest, v StreamVersion) (chan<- DeltaResponse, func()) {
+func (cache *MuxCache) CreateDeltaWatch(request *DeltaRequest, v StreamVersion) (chan DeltaResponse, func()) {
 	return nil, nil
 }